+++ src/penn_chime/charts.py
@@ -12,12 +12,15 @@ from .parameters import Parameters
 from .hc_actuals import ADMISSIONS_COLUMNS, CENSUS_COLUMNS
 
 
-def get_actual_columns_and_colors(possible_columns, projections_df, actuals_df, alt):
+def get_actual_columns_and_colors(possible_columns, projections_df, actuals_df, p, alt):
     actuals_plot_columns = []
     actuals_color_domain = []
-    actuals_colors_defaults = ['#4C78A8', '#F58518', '#E45756']
+    actuals_colors_defaults = ['black', '#4C78A8', '#F58518', '#E45756']
+    if "daily_regional_infections" in possible_columns:
+        actuals_colors_defaults = ['#4C78A8']
     actuals_color_range=[]
     combined = projections_df.merge(actuals_df, on="date", how="left")
+    # combined = combined.rename(columns=p.actuals_labels)
     for index, possible_column in enumerate(possible_columns):
         if possible_column in actuals_df.columns:
             actuals_plot_columns.append(possible_column)
@@ -26,29 +29,54 @@ def get_actual_columns_and_colors(possible_columns, projections_df, actuals_df, 
     actuals_color = alt.Color("Actual:N", scale=alt.Scale(domain=actuals_color_domain, range=actuals_color_range))
     return combined, actuals_plot_columns, actuals_color
         
+def get_proj_columns_and_colors(df, p, alt):
+    # departments
+    df = df.rename(columns={
+        "total": p.labels["total"],
+        "non_icu": p.labels["non_icu"],
+        "icu": p.labels["icu"],
+        "ventilators": p.labels["ventilators"],
+    })
+    plot_columns = [
+        p.labels["total"],
+        p.labels["non_icu"],
+        p.labels["icu"],
+        p.labels["ventilators"]
+    ]
+
+    proj_color = alt.Color(
+        "Projected:N",
+        sort=plot_columns,
+        scale=alt.Scale(
+            domain=plot_columns,
+            range=['black', '#4C78A8', '#F58518', '#E45756'],
+        )
+    )
+    return df, plot_columns, proj_color
 
 def build_admits_chart(
     *,
     alt,
     admits_floor_df: pd.DataFrame,
-    parameters: Parameters,
+    p: Parameters,
     actuals: Union[pd.DataFrame, None],
 ) -> Chart:
     """Build admits chart."""
     y_scale = alt.Scale()
-    if parameters.max_y_axis_set:
-        y_scale.domain = (0, parameters.max_y_axis)
+    if p.max_y_axis_set:
+        y_scale.domain = (0, p.max_y_axis)
         y_scale.clamp = True
 
-    plot_columns = ["total", "icu", "ventilators"]
+    admits_floor_df, plot_columns, proj_color = get_proj_columns_and_colors(
+        admits_floor_df, p, alt)
+
     x = dict(shorthand="date:T", title="Date", axis=alt.Axis(format=(DATE_FORMAT)))
     y = dict(shorthand="value:Q", title="Daily admissions", scale=y_scale)
-    color = alt.Color("Projected:N", sort = plot_columns)
     tooltip=[alt.Tooltip("utcmonthdate(date):O", title="Date", format=(DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f"), "Projected:N"]
     lines = (
         alt.Chart()
         .transform_fold(fold=plot_columns, as_=["Projected", "value"])
-        .encode(x=alt.X(**x), y=alt.Y(**y), color=color, tooltip=tooltip)
+        .encode(x=alt.X(**x), y=alt.Y(**y), color=proj_color, tooltip=tooltip)
         .mark_line()
     )
     bar = (
@@ -60,7 +88,7 @@ def build_admits_chart(
     charts = [lines, bar]
     if actuals is not None:
         admits_floor_df, actuals_plot_columns, actuals_color = get_actual_columns_and_colors(
-            ADMISSIONS_COLUMNS, admits_floor_df, actuals, alt,
+            ADMISSIONS_COLUMNS, admits_floor_df, actuals, p, alt,
         )
         actuals_tooltip=[alt.Tooltip("utcmonthdate(date):O", title="Date", format=(DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f"), "Actual:N"]
         actuals_lines = (
@@ -79,27 +107,27 @@ def build_census_chart(
     *,
     alt,
     census_floor_df: pd.DataFrame,
-    parameters: Parameters,
+    p: Parameters,
     actuals: Union[pd.DataFrame, None],
 ) -> Chart:
     """Build census chart."""
     y_scale = alt.Scale()
-    if parameters.max_y_axis_set:
-        y_scale.domain = (0, parameters.max_y_axis)
+    if p.max_y_axis_set:
+        y_scale.domain = (0, p.max_y_axis)
         y_scale.clamp = True
 
-    plot_columns = ["total", "icu", "ventilators"]
+    census_floor_df, plot_columns, proj_color = get_proj_columns_and_colors(
+        census_floor_df, p, alt)
+
     x = dict(shorthand="date:T", title="Date", axis=alt.Axis(format=(DATE_FORMAT)))
     y = dict(shorthand="value:Q", title="Census", scale=y_scale)
-    color = alt.Color("Projected:N", sort = plot_columns)
-    # tooltip = [alt.Tooltip("date:T", format=(DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f", title="Census"), "Projected:N"]
     tooltip = [alt.Tooltip("utcmonthdate(date):O", title="Date", format=(DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f", title="Census"), "Projected:N"]
 
     # TODO fix the fold to allow any number of dispositions
     lines = (
         alt.Chart()
         .transform_fold(fold=plot_columns, as_=["Projected", "value"])
-        .encode(x=alt.X(**x), y=alt.Y(**y), color=color, tooltip=tooltip)
+        .encode(x=alt.X(**x), y=alt.Y(**y), color=proj_color, tooltip=tooltip)
         .mark_line()
     )
     bar = (
@@ -111,7 +139,7 @@ def build_census_chart(
     charts = [lines, bar]
     if actuals is not None:
         census_floor_df, actuals_plot_columns, actuals_color = get_actual_columns_and_colors(
-            CENSUS_COLUMNS, census_floor_df, actuals, alt,
+            CENSUS_COLUMNS, census_floor_df, actuals, p, alt,
         )
         actuals_tooltip=[alt.Tooltip("utcmonthdate(date):O", title="Date", format=(DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f", title="Census"), "Actual:N"]
         actuals_lines = (
@@ -129,6 +157,7 @@ def build_sim_sir_w_date_chart(
     alt,
     sim_sir_w_date_floor_df: pd.DataFrame,
     actuals: Union[pd.DataFrame, None],
+    p: Parameters,
 ) -> Chart:
     """Build sim sir w date chart."""
     y_scale = alt.Scale()
@@ -155,7 +184,7 @@ def build_sim_sir_w_date_chart(
     charts = [lines, bar]
     if actuals is not None:
         sim_sir_w_date_floor_df, actuals_plot_columns, actuals_color = get_actual_columns_and_colors(
-            ["daily_regional_infections"], sim_sir_w_date_floor_df, actuals, alt,
+            ["daily_regional_infections"], sim_sir_w_date_floor_df, actuals, p, alt,
         )
         actuals_tooltip=[alt.Tooltip("utcmonthdate(date):O", title="Date", format=(DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f"), "Actual:N"]
         actuals_lines = (
@@ -178,18 +207,16 @@ def build_beds_chart(
         y_scale.domain = (-parameters.max_y_axis, parameters.max_y_axis)
         y_scale.clamp = True
 
-    beds_columns = ["total", "icu", "ventilators"]
+    beds_floor_df, beds_columns, proj_color = get_proj_columns_and_colors(
+        beds_floor_df, parameters, alt)
     x = dict(shorthand="date:T", title="Date", axis=alt.Axis(format=(DATE_FORMAT)))
     y = dict(shorthand="value:Q", title="COVID-19 Capacity", scale=y_scale)
-    color = alt.Color("Projected:N", sort = beds_columns)
     tooltip = ["Projected:N", alt.Tooltip("value:Q", format=".0f")]
     
-    beds_floor_df["zero"] = 0
-    # TODO fix the fold to allow any number of dispositions
     beds = (
         alt.Chart()
         .transform_fold(fold=beds_columns, as_=["Projected", "value"])
-        .encode(x=alt.X(**x), y=alt.Y(**y), color=color, tooltip=tooltip)
+        .encode(x=alt.X(**x), y=alt.Y(**y), color=proj_color, tooltip=tooltip)
         .mark_line()
     )
     bar = (
@@ -215,7 +242,7 @@ def build_ppe_chart(
     plot_columns: str,
 ) -> Chart:
     """Build ppe chart."""
-    k = list(p.ppe_labels.keys())[2:]
+    k = list(p.ppe_labels.keys())[3:]
     if plot_columns not in k:
         raise ValueError("PPE type must be in %s" % (k))
 
@@ -229,14 +256,27 @@ def build_ppe_chart(
     y_axis_label = "Required " + chart_title
     # departments
     ppe_floor_df = ppe_floor_df.rename(columns={
-        p.ppe_labels[plot_columns]["col1_name"]:p.ppe_labels["total"],
-        p.ppe_labels[plot_columns]["col2_name"]:p.ppe_labels["icu"],
-        })
-    plot_columns = [p.ppe_labels["total"], p.ppe_labels["icu"]]
+        p.ppe_labels[plot_columns]["col1_name"]: p.ppe_labels["total"],
+        p.ppe_labels[plot_columns]["col2_name"]: p.ppe_labels["non_icu"],
+        p.ppe_labels[plot_columns]["col3_name"]:p.ppe_labels["icu"],
+    })
+    plot_columns = [
+        p.ppe_labels["total"],
+        p.ppe_labels["non_icu"],
+        p.ppe_labels["icu"],
+    ]
+    # Set colors
+    proj_color = alt.Color(
+        "Department:N", 
+        sort=plot_columns,
+        scale = alt.Scale(
+            domain=plot_columns,
+            range=['black', '#4C78A8', '#F58518'],
+        )
+    )
     x = dict(shorthand="date:T", title="Date",
              axis=alt.Axis(format=(DATE_FORMAT)))
     y = dict(shorthand="value:Q", title=y_axis_label, scale=y_scale)
-    color = alt.Color("Department:N", sort=plot_columns)
     tooltip = [alt.Tooltip("utcmonthdate(date):O", title="Date", format=(
         DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f", title=chart_title), "Department:N"]
 
@@ -244,7 +284,7 @@ def build_ppe_chart(
     lines = (
         alt.Chart(title=chart_title)
         .transform_fold(fold=plot_columns, as_=["Department", "value"])
-        .encode(x=alt.X(**x), y=alt.Y(**y), color=color, tooltip=tooltip)
+        .encode(x=alt.X(**x), y=alt.Y(**y), color=proj_color, tooltip=tooltip)
         .mark_line()
     )
     bar = (
@@ -282,18 +322,27 @@ def build_staffing_chart(
     y_axis_label = "Required " + chart_title
     # departments
     staffing_floor_df = staffing_floor_df.rename(columns={
-        p.staffing_labels[plot_columns]["col3_name"]: p.staffing_labels["total"],
-        p.staffing_labels[plot_columns]["col2_name"]: p.staffing_labels["icu"],
-        p.staffing_labels[plot_columns]["col1_name"]: p.staffing_labels["nonicu"],
+        p.staffing_labels[plot_columns]["col1_name"]: p.staffing_labels["total"],
+        p.staffing_labels[plot_columns]["col2_name"]: p.staffing_labels["non_icu"],
+        p.staffing_labels[plot_columns]["col3_name"]: p.staffing_labels["icu"],
     })
     plot_columns = [
         p.staffing_labels["total"], 
+        p.staffing_labels["non_icu"],
         p.staffing_labels["icu"],
-    ]
+    ] 
+    # Set colors
+    proj_color = alt.Color(
+        "Department:N",
+        sort=plot_columns,
+        scale=alt.Scale(
+            domain=plot_columns,
+            range=['black', '#4C78A8', '#F58518'],
+        )
+    )
     x = dict(shorthand="date:T", title="Date",
              axis=alt.Axis(format=(DATE_FORMAT)))
     y = dict(shorthand="value:Q", title=y_axis_label, scale=y_scale)
-    color = alt.Color("Department:N", sort=plot_columns)
     tooltip = [alt.Tooltip("utcmonthdate(date):O", title="Date", format=(
         DATE_FORMAT)), alt.Tooltip("value:Q", format=".0f", title=chart_title), "Department:N"]
 
@@ -301,7 +350,7 @@ def build_staffing_chart(
     lines = (
         alt.Chart(title=chart_title)
         .transform_fold(fold=plot_columns, as_=["Department", "value"])
-        .encode(x=alt.X(**x), y=alt.Y(**y), color=color, tooltip=tooltip)
+        .encode(x=alt.X(**x), y=alt.Y(**y), color=proj_color, tooltip=tooltip)
         .mark_line()
     )
     bar = (
@@ -320,8 +369,7 @@ def build_staffing_chart(
 def build_descriptions(
     *,
     chart: Chart,
-    labels: Dict[str, str],
-    suffix: str = ""
+    labels: dict,
 ) -> str:
     """
 
@@ -333,7 +381,7 @@ def build_descriptions(
     """
     messages = []
 
-    cols = ["total", "icu", "ventilators"]
+    cols = ["Total", "Non-ICU", "ICU", "Ventilators"]
     asterisk = False
     day = "date" if "date" in chart.data.columns else "day"
 
@@ -360,8 +408,7 @@ def build_descriptions(
 def build_bed_descriptions(
     *,
     chart: Chart,
-    labels: Dict[str, str],
-    suffix: str = "",
+    labels: dict,
 ) -> str:
     """
 
@@ -373,11 +420,11 @@ def build_bed_descriptions(
     """
     messages = []
 
-    cols = ["total", "icu", "ventilators"]
+    cols = ["Total", "Non-ICU", "ICU", "Ventilators"]
     asterisk = False
     
     # Add note if lines overlap.
-    if sum(np.where(chart.data["total"] == chart.data["icu"], 1, 0)) > 1:
+    if sum(np.where(chart.data["Total"] == chart.data["ICU"], 1, 0)) > 1:
         messages.append("_The overlapping lines represent non-ICU patients being housed in the ICU._")
 
     for col in cols:
@@ -408,7 +455,7 @@ def build_ppe_descriptions(
     """
     messages = []
 
-    cols = ["Total", "ICU"]
+    cols = ["Total", "Non-ICU", "ICU"]
     asterisk = False
     day = "date" if "date" in chart.data.columns else "day"
 
@@ -445,7 +492,7 @@ def build_staffing_descriptions(
     """
     messages = []
 
-    cols = ["Total", "ICU"]
+    cols = ["Total", "Non-ICU", "ICU"]
     asterisk = False
     day = "date" if "date" in chart.data.columns else "day"
 
@@ -488,22 +535,22 @@ def build_table(
             "day",
             "date",
             "masks_n95_total",
-            "masks_n95_hosp",
+            "masks_n95_non_icu",
             "masks_n95_icu",
             "masks_surgical_total",
-            "masks_surgical_hosp",
+            "masks_surgical_non_icu",
             "masks_surgical_icu",
             "face_shield_total",
-            "face_shield_hosp",
+            "face_shield_non_icu",
             "face_shield_icu",
             "gloves_total",
-            "gloves_hosp",
+            "gloves_non_icu",
             "gloves_icu",
             "gowns_total",
-            "gowns_hosp",
+            "gowns_non_icu",
             "gowns_icu",
             "other_ppe_total",
-            "other_ppe_hosp",
+            "other_ppe_non_icu",
             "other_ppe_icu",
         ]]
     if "nurses_total" in table_df.columns:
@@ -511,16 +558,16 @@ def build_table(
             "day",
             "date",
             "nurses_total",
-            "nurses_hosp",
+            "nurses_non_icu",
             "nurses_icu",
             "physicians_total",
-            "physicians_hosp",
+            "physicians_non_icu",
             "physicians_icu",
             "advanced_practice_providers_total",
-            "advanced_practice_providers_hosp",
+            "advanced_practice_providers_non_icu",
             "advanced_practice_providers_icu",
             "healthcare_assistants_total",
-            "healthcare_assistants_hosp",
+            "healthcare_assistants_non_icu",
             "healthcare_assistants_icu",
         ]]
     return table_df