+++ changelog.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Non-ICU is now displayed on all charts. 
+### Changed
+- "Hospitalized" was renamed to "Non-ICU" across the whole app.
+- Total is now black, non-icu is blue. This matches penn.
 
 ## [1.6.0] - 2020-04-13
 ### Added